+++ app/assets/javascripts/shopify_app/request_storage_access.js
@@ -1,0 +1,3 @@
+//= require ./itp_helper.js
+//= require ./storage_access.js
+//= require ./storage_access_redirect.js