+++ lib/generators/shopify_app/install/install_generator.rb
@@ -9,16 +9,24 @@ module ShopifyApp
       class_option :application_name, type: :array, default: ['My', 'Shopify', 'App']
       class_option :api_key, type: :string, default: '<api_key>'
       class_option :secret, type: :string, default: '<secret>'
+      class_option :old_secret, type: :string, default: '<old_secret>'
       class_option :scope, type: :array, default: ['read_products']
       class_option :embedded, type: :string, default: 'true'
+      class_option :api_version, type: :string, default: ShopifyAPI::ApiVersion.latest_stable_version.to_s
 
       def create_shopify_app_initializer
         @application_name = format_array_argument(options['application_name'])
         @api_key = options['api_key']
         @secret = options['secret']
+        @old_secret = options['old_secret']
         @scope = format_array_argument(options['scope'])
+        @api_version = options['api_version']
 
         template 'shopify_app.rb', 'config/initializers/shopify_app.rb'
+      end
+
+      def create_session_store_initializer
+        copy_file('session_store.rb', 'config/initializers/session_store.rb')
       end
 
       def create_and_inject_into_omniauth_initializer