+++ lib/shopify_app/configuration.rb
@@ -7,6 +7,7 @@ module ShopifyApp
     attr_accessor :application_name
     attr_accessor :api_key
     attr_accessor :secret
+    attr_accessor :old_secret
     attr_accessor :scope
     attr_accessor :embedded_app
     alias_method  :embedded_app?, :embedded_app
@@ -14,6 +15,7 @@ module ShopifyApp
     attr_accessor :scripttags
     attr_accessor :after_authenticate_job
     attr_accessor :session_repository
+    attr_accessor :api_version
 
     # customise urls
     attr_accessor :root_url