+++ test/generators/install_generator_test.rb
@@ -12,26 +12,36 @@ class InstallGeneratorTest < Rails::Generators::TestCase
     provide_existing_application_controller
   end
 
+  teardown do
+    ShopifyAPI::ApiVersion.clear_defined_versions
+    ShopifyAPI::ApiVersion.define_known_versions
+  end
+
   test "creates the ShopifyApp initializer" do
+    latest_stable_version = ShopifyAPI::ApiVersion.latest_stable_version
+
     run_generator
     assert_file "config/initializers/shopify_app.rb" do |shopify_app|
       assert_match 'config.application_name = "My Shopify App"', shopify_app
       assert_match 'config.api_key = "<api_key>"', shopify_app
       assert_match 'config.secret = "<secret>"', shopify_app
+      assert_match 'config.old_secret = "<old_secret>"', shopify_app
       assert_match 'config.scope = "read_products"', shopify_app
       assert_match "config.embedded_app = true", shopify_app
+      assert_match "config.api_version = \"#{latest_stable_version}\"", shopify_app
       assert_match "config.after_authenticate_job = false", shopify_app
     end
   end
 
   test "creates the ShopifyApp initializer with args" do
-    run_generator %w(--application_name Test Name --api_key key --secret shhhhh --scope read_orders, write_products)
+    run_generator %w(--application_name Test Name --api_key key --secret shhhhh --api_version unstable --scope read_orders, write_products)
     assert_file "config/initializers/shopify_app.rb" do |shopify_app|
       assert_match 'config.application_name = "Test Name"', shopify_app
       assert_match 'config.api_key = "key"', shopify_app
       assert_match 'config.secret = "shhhhh"', shopify_app
       assert_match 'config.scope = "read_orders, write_products"', shopify_app
       assert_match 'config.embedded_app = true', shopify_app
+      assert_match 'config.api_version = "unstable"', shopify_app
       assert_match 'config.session_repository = ShopifyApp::InMemorySessionStore', shopify_app
     end
   end